updatecart value function if user refresh the page it will remain same as the length of array in localstorage.

+++ addToCart.js
@@ -1,20 +1,29 @@
 import { getCartProductFromLS } from "./getCartProductFromLS.JS";
+import { updateCartValue } from "./updateCartValue";
 
-export const addToCart=(event,id,stock)=>{
+getCartProductFromLS();
 
-    let arrLocalStorageProduct=getCartProductFromLS();
-    const currentElement=document.querySelector(`#card${id}`);
+export const addToCart = (event, id, stock) => {
+  let arrLocalStorageProduct = getCartProductFromLS();
+  const currentElement = document.querySelector(`#card${id}`);
 
-    let quantity=currentElement.querySelector('.productQuantity').innerText;
-    let price=currentElement.querySelector('.productPrice').innerText;
+  let quantity = currentElement.querySelector(".productQuantity").innerText;
+  let price = currentElement.querySelector(".productPrice").innerText;
 
-    // console.log(price,quantity);
+  console.log(price, quantity);
 
-    price=price.replace("$","");
-    price=Number(price*quantity);
-    quantity=Number(quantity);
-    // console.log(price,quantity);
+  price = price.replace("$", "");
 
-    arrLocalStorageProduct.push({id,quantity,price});
-    localStorage.setItem('cartProductLS',JSON.stringify(arrLocalStorageProduct));
+  let existingProd = arrLocalStorageProduct.find((curProd) => curProd.id == id);
+  if (existingProd) {
+    return false;
+  }
+  price = Number(price * quantity);
+  quantity = Number(quantity);
+  console.log(price, quantity);
+
+  arrLocalStorageProduct.push({ id, quantity, price });
+  localStorage.setItem("cartProductLS", JSON.stringify(arrLocalStorageProduct));
+
+  updateCartValue(arrLocalStorageProduct);
 };
